+++ packages/liblsl/test/liblsl_stream_info_metadata_test.dart
@@ -343,7 +343,8 @@ void main() {
         // For some unknown reason this test just fails on CI
         // TODO: Investigate why this test is flaky in CI but works on
         // OSX / Win / Linux locally
-        skip: true,
+        skip: false,
+        tags: ['flaky'],
       );
     });